szabolcs_rajcsanyi/qna/requirements.txt
langchain==0.1.16
langchain_openai==0.1.4
streamlit==1.27.2
python-dotenv==0.19.2
pypdf==3.17.0
faiss-cpu==1.7.4